moved route logic into an api folder
