+++ server.js
@@ -4,40 +4,27 @@ const app = express();
 //create server port
 const PORT = process.env.PORT || 3030;
 //retrieve the data from my created route
-const triviaData = require('./routes/trivia');
+// const triviaData = require('./api/trivia');
 // const path = require('path');
 
 // //static files
 // app.use(express.static(path.join(__dirname, 'build/')));
 
 //use the express Router object to create our route
-let router = express.Router();
+// let router = express.Router();
 
 //configure middleware to support JSON data parsing from client requests
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
+//configure router so routes are prefixed with /api
+app.use('/api', require('./api'));
+
 if (process.env.NODE_ENV && process.env.NODE_ENV !== 'development') {
-  app.get('*', (req, res) => {
+  app.get('/', (req, res) => {
     res.sendFile('build/index.html', { root: __dirname });
   });
 }
-
-//configure router so routes are prefixed with /api
-app.use('/api', router);
-
-//create a GET request to gather all the questions
-router.get('/', (req, res, next) => {
-  //use the get method from the routes/trivia.js
-  triviaData.get(
-    function (data) {
-      res.status(200).json(data);
-    },
-    function (err) {
-      next(err);
-    }
-  );
-});
 
 //configure error handling middleware last
 app.use(function (err, req, res, next) {
